backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
msal==1.26.0
openai==1.3.9
requests==2.31.0
azure-identity==1.15.0
azure-storage-blob==12.19.0